+++ html5lib/tests/test_serializer.py
@@ -74,7 +74,7 @@ class JsonWalker(TreeWalker):
         attrs = {}
         for attrib in attribs:
             name = (attrib["namespace"], attrib["name"])
-            assert(name not in attrs)
+            assert name not in attrs
             attrs[name] = attrib["value"]
         return attrs